Refactor database connection handling and executeStoredProcedure function
